requirements.txt
qiskit>=0.34.0
tensorflow>=2.7.0
numpy>=1.21.0
pytorch>=1.10.0
pytest>=6.2.5
pyyaml>=5.4.1
matplotlib>=3.4.3
pandas>=1.3.4
